Make poster work with playlists and quality selector
Must deal with these conditions:
- must be set up as player event outside the engine as the first clip
  may be not HLS
- but a HLS clip may be played first by clicking on a playlist trigger
  element
- with playlists there is no player event where you can catch the poster
  class reliably

For the moment we do not confine this to the hlsjs engine, as it works
around most of the problems in
https://github.com/flowplayer/flowplayer/issues/942
if kept unconditionally. If the minimal timeout happens outside the
hlsjs engine it serves as harmles double check for poster state, but
sets it if missed by core/quality selector.