Re-apply autoplay nudge to all clients (#97)
As a side effect this works around flowplayer/flowplayer#1260 if the
stream can be played with hlsjs: the user will be prompted to start
playback to overcome the broken play promise.

Also revert to omitting video tag autoplay attribute which is, at least
in theory, not applicable for MediaSource. This seems to remove some
extra nudges.

Moving nudge to ready breaks in some Android clients.